requirements.txt
psutil==5.9.8
requests==2.32.4
fastapi==0.115.13
uvicorn==0.34.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2 